+++ lib/wechat/api_base.rb
@@ -16,12 +16,20 @@ module Wechat
       get 'media/get', params: { media_id: media_id }, as: :file
     end
 
+    def media_hq(media_id)
+      get 'media/get/jssdk', params: { media_id: media_id }, as: :file
+    end
+
     def media_create(type, file)
       post_file 'media/upload', file, params: { type: type }
     end
 
     def media_uploadimg(file)
       post_file 'media/uploadimg', file
+    end
+
+    def media_uploadnews(mpnews_message)
+      post 'media/uploadnews', mpnews_message.to_json
     end
 
     protected