+++ lib/wechat/controller_api.rb
@@ -3,61 +3,94 @@ module Wechat
     extend ActiveSupport::Concern
 
     module ClassMethods
-      attr_accessor :wechat_api_client, :token, :appid, :corpid, :agentid, :encrypt_mode, :timeout,
+      attr_accessor :wechat_api_client, :wechat_cfg_account, :token, :appid, :corpid, :agentid, :encrypt_mode, :timeout,
                     :skip_verify_ssl, :encoding_aes_key, :trusted_domain_fullname, :oauth2_cookie_duration
     end
 
-    def wechat
-      self.class.wechat # Make sure user can continue access wechat at instance level similar to class level
+    def wechat(account = nil)
+      # Make sure user can continue access wechat at instance level similar to class level
+      self.class.wechat(account)
     end
 
-    def wechat_oauth2(scope = 'snsapi_base', page_url = nil, &block)
-      appid = self.class.corpid || self.class.appid
-      page_url ||= if self.class.trusted_domain_fullname
-                     "#{self.class.trusted_domain_fullname}#{request.original_fullpath}"
-                   else
-                     request.original_url
-                   end
-      redirect_uri = CGI.escape(page_url)
-      oauth2_url = "https://open.weixin.qq.com/connect/oauth2/authorize?appid=#{appid}&redirect_uri=#{redirect_uri}&response_type=code&scope=#{scope}&state=#{wechat.jsapi_ticket.oauth2_state}#wechat_redirect"
+    def wechat_oauth2(scope = 'snsapi_base', page_url = nil, account = nil, &block)
+      # ensure wechat initialization
+      self.class.corpid || self.class.appid || self.class.wechat
 
-      return oauth2_url unless block_given?
-      if self.class.corpid
-        wechat_corp_oauth2(oauth2_url, &block)
+      api = wechat(account)
+      if account
+        config = Wechat.config(account)
+        appid = config.corpid || config.appid
+        is_crop_account = !!config.corpid
       else
-        wechat_public_oauth2(oauth2_url, &block)
+        appid = self.class.corpid || self.class.appid
+        is_crop_account = !!self.class.corpid
       end
+
+      raise 'Can not get corpid or appid, so please configure it first to using wechat_oauth2' if appid.blank?
+
+      api.jsapi_ticket.ticket if api.jsapi_ticket.oauth2_state.nil?
+      oauth2_params = {
+        appid: appid,
+        redirect_uri: page_url || generate_redirect_uri(account),
+        scope: scope,
+        response_type: 'code',
+        state: api.jsapi_ticket.oauth2_state
+      }
+
+      return generate_oauth2_url(oauth2_params) unless block_given?
+      is_crop_account ? wechat_corp_oauth2(oauth2_params, account, &block) : wechat_public_oauth2(oauth2_params, account, &block)
     end
 
     private
 
-    def wechat_public_oauth2(oauth2_url)
-      if cookies.signed_or_encrypted[:we_openid].blank? && params[:code].blank?
-        redirect_to oauth2_url
-      elsif cookies.signed_or_encrypted[:we_openid].blank? &&
-            params[:code].present? &&
-            params[:state].to_s == wechat.jsapi_ticket.oauth2_state.to_s # params[:state] maybe '' and wechat.jsapi_ticket.oauth2_state may be nil
-        access_info = wechat.web_access_token(params[:code])
+    def wechat_public_oauth2(oauth2_params, account = nil)
+      openid  = cookies.signed_or_encrypted[:we_openid]
+      unionid = cookies.signed_or_encrypted[:we_unionid]
+      we_token = cookies.signed_or_encrypted[:we_access_token]
+      if openid.present?
+        yield openid, { 'openid' => openid, 'unionid' => unionid, 'access_token' => we_token}
+      elsif params[:code].present? && params[:state] == oauth2_params[:state]
+        access_info = wechat(account).web_access_token(params[:code])
         cookies.signed_or_encrypted[:we_openid] = { value: access_info['openid'], expires: self.class.oauth2_cookie_duration.from_now }
+        cookies.signed_or_encrypted[:we_unionid] = { value: access_info['unionid'], expires: self.class.oauth2_cookie_duration.from_now }
+        cookies.signed_or_encrypted[:we_access_token] = { value: access_info['access_token'], expires: self.class.oauth2_cookie_duration.from_now }
         yield access_info['openid'], access_info
       else
-        yield cookies.signed_or_encrypted[:we_openid], { 'openid' => cookies.signed_or_encrypted[:we_openid] }
+        redirect_to generate_oauth2_url(oauth2_params)
       end
     end
 
-    def wechat_corp_oauth2(oauth2_url)
-      if cookies.signed_or_encrypted[:we_deviceid].blank? && params[:code].blank?
-        redirect_to oauth2_url
-      elsif cookies.signed_or_encrypted[:we_deviceid].blank? && params[:code].present? && params[:state] == wechat.jsapi_ticket.oauth2_state
-        userinfo = wechat.getuserinfo(params[:code])
+    def wechat_corp_oauth2(oauth2_params, account = nil)
+      userid   = cookies.signed_or_encrypted[:we_userid]
+      deviceid = cookies.signed_or_encrypted[:we_deviceid]
+      if userid.present? && deviceid.present?
+        yield userid, { 'UserId' => userid, 'DeviceId' => deviceid }
+      elsif params[:code].present? && params[:state] == oauth2_params[:state]
+        userinfo = wechat(account).getuserinfo(params[:code])
         cookies.signed_or_encrypted[:we_userid] = { value: userinfo['UserId'], expires: self.class.oauth2_cookie_duration.from_now }
         cookies.signed_or_encrypted[:we_deviceid] = { value: userinfo['DeviceId'], expires: self.class.oauth2_cookie_duration.from_now }
-        cookies.signed_or_encrypted[:we_openid] = { value: userinfo['OpenId'], expires: self.class.oauth2_cookie_duration.from_now }
         yield userinfo['UserId'], userinfo
       else
-        yield cookies.signed_or_encrypted[:we_userid], { 'UserId' => cookies.signed_or_encrypted[:we_userid],
-                                                         'DeviceId' => cookies.signed_or_encrypted[:we_deviceid],
-                                                         'OpenId' => cookies.signed_or_encrypted[:we_openid] }
+        redirect_to generate_oauth2_url(oauth2_params)
+      end
+    end
+
+    def generate_redirect_uri(account = nil)
+      domain_name = if account
+        Wechat.config(account).trusted_domain_fullname
+      else
+        self.class.trusted_domain_fullname
+      end
+      page_url = domain_name ? "#{domain_name}#{request.original_fullpath}" : request.original_url
+      safe_query = request.query_parameters.reject { |k, _| %w(code state access_token).include? k }.to_query
+      page_url.sub(request.query_string, safe_query)
+    end
+
+    def generate_oauth2_url(oauth2_params)
+      if oauth2_params[:scope] == 'snsapi_login'
+        "https://open.weixin.qq.com/connect/qrconnect?#{oauth2_params.to_query}#wechat_redirect"
+      else
+        "https://open.weixin.qq.com/connect/oauth2/authorize?#{oauth2_params.to_query}#wechat_redirect"
       end
     end
   end