+++ spec/lib/wechat/helpers_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-class WechatApiController < ActionController::Base
+class WechatApiController < ActionController::API
 end
 
 RSpec.describe WechatApiController, type: :controller do
@@ -15,11 +15,35 @@ RSpec.describe WechatApiController, type: :controller do
     end
 
     it '#wechat_config_js' do
-      controller.request = ActionController::TestRequest.create
+      controller.request = ActionController::TestRequest.create ActionController::TestRequest::DEFAULT_ENV
       controller.request.host = 'test.host'
       expect(controller.wechat.jsapi_ticket).to receive(:signature)
         .with('http://test.host').and_return(js_hash_result)
       expect(wechat_config_js(debug: false, api: %w(hideMenuItems openEnterpriseChat))).to end_with '</script>'
+    end
+  end
+
+  describe '#wechat_config_js with account' do
+    before(:all) do
+      Wechat::ApiLoader.class_eval { @configs = nil }
+      ENV['WECHAT_CONF_FILE'] = File.join(Dir.getwd, 'spec/dummy/config/dummy_wechat.yml')
+    end
+
+    after(:all) do
+      Wechat::ApiLoader.class_eval { @configs = nil }
+      ENV['WECHAT_CONF_FILE'] = nil
+    end
+
+    controller do
+      wechat_api
+    end
+
+    it '#wechat_config_js' do
+      controller.request = ActionController::TestRequest.create ActionController::TestRequest::DEFAULT_ENV
+      controller.request.host = 'test.host'
+      expect(Wechat.api(:wx2).jsapi_ticket).to receive(:signature)
+        .with('http://test.host').and_return(js_hash_result)
+      expect(wechat_config_js(account: :wx2, debug: false, api: %w(hideMenuItems))).to end_with '</script>'
     end
   end
 
@@ -29,11 +53,12 @@ RSpec.describe WechatApiController, type: :controller do
     end
 
     it 'called with trusted_domain' do
-      controller.request = ActionController::TestRequest.create
+      controller.request = ActionController::TestRequest.create ActionController::TestRequest::DEFAULT_ENV
       controller.request.host = 'test.host'
       expect(controller.wechat.jsapi_ticket).to receive(:signature)
         .with('http://trusted.host').and_return(js_hash_result)
       expect(wechat_config_js(debug: false, api: %w(hideMenuItems))).to end_with '</script>'
     end
   end
+
 end